requirements.txt
Flask==1.0.2
google-api-core==1.7.0
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-cloud-speech==0.36.3
google-cloud-texttospeech==0.4.0
googleapis-common-protos==1.5.6
grpcio==1.18.0
gunicorn==19.9.0
numpy==1.16.0
requests==2.21.0
Werkzeug==0.14.1
